Allow proposal editing via a slug.
We'll give submitters a slugged url to edit their proposal. We'll need to warn them
to not share or lose their slug, and let them know when editing ends.

This commit also cleans up the routes a little bit and makes it possible to reuse the
collection form for editing. I'm not very happy with the update method in the collect controller,
but it'll do until someone comes up with better.