Implements a CanCan like mixin that utilizes static classes for access
This is controller based security and unlike CanCan doesn't care about the models

Right now this only handles things on a controller level and I don't plan on extending to the actions but we could totally do so

Its a lot of meta programming to get around the class read vs singleton presence time in ruby otherwise its pretty simple and declarative in the user_cans class

Order is dependant for the authenticate! before filter and the include HasAccess.

This allows a user to autosign even if they do not have access to a controller. Just dumps them back to the homepage