Add sorting javascript
This is sort of horrible because I can't figure out how to get this
library to turn off sorting based on a class, so all columns are
sortable but I'm only displaying a diamond for the .sortable headers.

Also it doesn't cycle through "turn sorting off" which pisses me off.