Fixes #6. Repairs confirmation step for proposal destroy.
Dunno if this was an upgrade issue or what, but `link_to` wasn't constructing the `data-confirm` attribute in the anchor right, so the ujs wasn't picking it up when the link was activated. So I'm using ugly symbols now, but the task was to fix the destroy confirmation, not rewrite the view.

I really wanna rewrite all the views. :(