Look, I'm not proud of this code.
It's, like, O(n^infinity). We're not using this version of blinder ever again. We've got to adapt the data model to make it reliable, secure, and easy to retrieve specific elements from the submitted proposals. Like the friggin' title of the proposal. Ugh.

Anyway, this will get us over the hump maybe.